fix(workflow): security hardening — tenant auth, SSRF, SQL injection, input guards
- Add tenant authorization checks to WorkflowLifecycleController and
  WorkflowConfigPanel to prevent cross-tenant access
- Enhance PreventsSSRF with IPv6 support and cloud metadata endpoint blocking
- Add field name validation in FindRecordAction and ScheduledTrigger to
  prevent SQL injection via dynamic column names
- Scope ScheduledTrigger queries to tenant to prevent cross-tenant execution
- Add tenant-scoped record resolution in DeleteRecord and UpdateRecord
  actions with model re-hydration from serialized context
- Cast DelayAction duration to int, wrap CreateRecordAction in try-catch

+++ packages/workflow/src/Actions/DelayAction.php
@@ -18,7 +18,7 @@ class DelayAction extends BaseAction
      */
     public function execute(array $config, array $context): array
     {
-        $duration = $config['duration'] ?? 0;
+        $duration = (int) ($config['duration'] ?? 0);
         $unit = $config['unit'] ?? 'minutes';
 
         return [
